Refactored /warehouses/new form into WarehouseForm
