+++ UI/config/routeRoleMapping.js
@@ -4,12 +4,16 @@ const all = ["manager", "clerk", "salesperson"];
 const managerOnly = ["manager"];
 
 // Item mappings
-roleRouteMappings.set("/items/new", managerOnly);
 roleRouteMappings.set("/items", all);
+roleRouteMappings.set("/items/new", managerOnly);
 roleRouteMappings.set("/items/:uuid", all);
 
 // User mappings
-roleRouteMappings.set("/users/new", managerOnly);
 roleRouteMappings.set("/users", managerOnly);
+roleRouteMappings.set("/users/new", managerOnly);
+
+// Warehouse mappings
+roleRouteMappings.set("/warehouses", all);
+roleRouteMappings.set("/warehouses/new", managerOnly);
 
 export default roleRouteMappings;
